server/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt[crypto]==5.5.1
gunicorn==25.3.0
packaging==26.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
